rewrite Makefile
PACKAGENAME を DEBIAN/control から取得して make install ; make uninstall を出来るようにした。

+++ Makefile
@@ -1,19 +1,59 @@
 
-.PHONY: all clean deb-package
+PACKAGENAME:=$(shell grep '^Package:' DEBIAN/control | sed 's/^Package:[[:space:]]*//')
+PACKAGEVERSION:=$(shell grep '^Version:' DEBIAN/control | sed 's/^Version:[[:space:]]*//')
+PACKAGEARCH:=$(shell grep '^Architecture:' DEBIAN/control | sed 's/^Architecture:[[:space:]]*//')
+PACKAGEFILE:=$(PACKAGENAME)_$(PACKAGEVERSION)_$(PACKAGEARCH).deb
+
+BUILD:=.build
+DEB_LIBEXEC:=$(BUILD)/usr/libexec/$(PACKAGENAME)
+DEB_SYSTEMD_SYSTEM:=$(BUILD)/lib/systemd/system
+
+DEB_LIBEXEC_COMPONENT:=seqtask.js pseudo-job
+DEB_SYSTEMD_SYSTEM_COMPONENT:=control postinst postrm
+DEB_SYSTEMD_SYSTEM_COMPONENT_PATH:=$(shell for f in $(DEB_SYSTEMD_SYSTEM_COMPONENT) ; do echo DEBIAN/$$f ; done )
+
+.PHONY: all clean deb-package install uninstall update status 
 
 all: deb-package
 
-deb-package:
-	if [ ! -d .build ] ; then mkdir .build ; fi
-	if [ ! -d .build/usr/libexec/node-seqexec ] ; then mkdir -p .build/usr/libexec/node-seqexec ; fi
-	if [ ! -d .build/lib/systemd/system ] ; then mkdir -p .build/lib/systemd/system ; fi
-	cp ./seqtask.js .build/usr/libexec/node-seqexec/
-	cp ./pseudo-job .build/usr/libexec/node-seqexec/
+$(BUILD):
+	if [ ! -d "$(BUILD)" ] ; then mkdir -p "$(BUILD)" ; fi
+
+$(DEB_LIBEXEC): $(BUILD) $(DEB_LIBEXEC_COMPONENT)
+	if [ ! -d "${DEB_LIBEXEC}" ] ; then mkdir -p "${DEB_LIBEXEC}" ; fi
+	for f in $(DEB_LIBEXEC_COMPONENT) ; do cp "$$f" "${DEB_LIBEXEC}" ; done
+
+$(DEB_SYSTEMD_SYSTEM): $(BUILD)
+	if [ ! -d "$(DEB_SYSTEMD_SYSTEM)" ] ; then mkdir -p "$(DEB_SYSTEMD_SYSTEM)" ; fi
 	cp ./node-seqexec.service .build/lib/systemd/system/ 
-	cp -r DEBIAN ./.build/DEBIAN 
+
+$(PACKAGEFILE): deb-package
+
+deb-package:$(DEB_SYSTEMD_SYSTEM) $(DEB_LIBEXEC) $(DEB_SYSTEMD_SYSTEM_COMPONENT_PATH)
+	if [ ! -d "$(BUILD)/DEBIAN" ] ; then mkdir -p "$(BUILD)/DEBIAN" ; fi
+	for f in $(DEB_SYSTEMD_SYSTEM_COMPONENT_PATH) ; do cp $$f $(BUILD)/DEBIAN ; done 
 	fakeroot dpkg-deb --build .build .
 
 clean:
 	find . -type f -name '*~' -delete
-	rm -rf .build
+	if [ -d $(BUILD) ]; then rm -rf .build ; fi 
+
+uninstall:
+	@if [ "installed" = "$$(if dpkg-query -s $(PACKAGENAME) > /dev/null 2>&1 ; then echo installed ; fi)" ] ; then sudo apt purge -y $(PACKAGENAME) ; fi
+
+install: $(PACKAGEFILE) uninstall
+	sudo apt install -y ./node-seqexec_0.0.0_all.deb
+
+update: $(PACKAGEFILE) 
+	make uninstall && make install
+
+status:
+
+check:
+	@echo $(DEB_SYSTEMD_SYSTEM_COMPONENT_PATH)
+	@echo "package name	: \"$(PACKAGENAME)\""
+	@echo "package version	: \"$(PACKAGEVERSION)\""
+	@echo "package file     : \"$(PACKAGEFILE)\""
+	@echo "libexec path	: \"$(DEB_LIBEXEC)\""
+	@echo "systemd_path	: \"$(DEB_SYSTEMD_SYSTEM)\""
